requirements.txt
crewai
crewai_tools
langchain_google_genai
load_dotenv